Fix changing size of input fielda and overflow of forms
there is no scroll on signup page anymore and the validation texts won't change the size of form anymore. Also the directory structure modified according to what we discussed in standup meeting.
