+++ client/src/pages/AuthRoutes/pages/Signup.js
@@ -9,11 +9,11 @@ import {
 } from '@material-ui/core';
 import {Formik, Form} from 'formik';
 import * as Yup from 'yup';
-import useStyles from '../styles/Auth.style';
-import Banner from '../components/Banner/components/Banner';
-import FormikControl from '../components/Formik/components/FormikControl';
+import useStyles from './Auth.style';
+import Banner from '../components/Banner';
+import FormikControl from '../../shared/Formik/FormikControl';
 
-const Login = () => {
+const Signup = () => {
 	const classes = useStyles();
 
 	const [open, setOpen] = useState(false);
@@ -23,17 +23,23 @@ const Login = () => {
 	};
 
 	const initialValues = {
+		name: '',
 		email: '',
 		password: '',
+		confirmPassword: '',
 	};
 
 	const validationSchema = Yup.object({
+		name: Yup.string().required('Name is Required'),
 		email: Yup.string()
 			.email('Invalid email format')
 			.required('Email is Required'),
 		password: Yup.string()
 			.required('Password is Required')
 			.min(6, 'Password too short'),
+		confirmPassword: Yup.string()
+			.oneOf([Yup.ref('password'), ''], 'Passwords must match')
+			.required('Confirm password is required'),
 	});
 
 	const onSubmit = (values) => {
@@ -47,7 +53,9 @@ const Login = () => {
 			<Grid item xs={12} sm={8} md={5} component={Paper} elevation={6} square>
 				<img src='/images/logo.jpg' alt='logo' className={classes.margin} />
 				<div className={classes.paper}>
-					<Typography variant='h3'>Login</Typography>
+					<Typography component='h3' variant='h3'>
+						Create an account
+					</Typography>
 					<div className={classes.form}>
 						<Formik
 							initialValues={initialValues}
@@ -57,22 +65,38 @@ const Login = () => {
 								<Form>
 									<FormikControl
 										control='input'
+										type='name'
+										label='Name'
+										name='name'
+										placeholder='Enter your name'
+									/>
+									<FormikControl
+										control='input'
 										type='email'
 										label='EMAIL'
 										name='email'
+										placeholder='Enter your e-mail address'
 									/>
 									<FormikControl
 										control='input'
 										type='password'
 										label='PASSWORD'
 										name='password'
+										placeholder='Enter password'
+									/>
+									<FormikControl
+										control='input'
+										type='password'
+										label='CONFIRM PASSWORD'
+										name='confirmPassword'
+										placeholder='Confirm password'
 									/>
 									<Button
 										type='submit'
 										variant='contained'
 										color='secondary'
 										className={classes.btn}>
-										Sign In
+										Sign Up
 									</Button>
 								</Form>
 							)}
@@ -91,9 +115,9 @@ const Login = () => {
 					}}
 				/>
 			</Grid>
-			<Banner />
+			<Banner signUp={true} />
 		</Grid>
 	);
 };
 
-export default Login;
+export default Signup;
